requirements.txt
flask
gunicorn
PyPDF2
sentence-transformers
faiss-cpu
huggingface_hub
python-dotenv
